The arrow file format is basically the streaming format plus a
flatbuffer footer with offset information to allow indexing. We could
pull an abstraction for reading and writing the streaming format out of
this if we need to, but I didn't do it here to save the complexity until
we need it.

I'm hoping to use this to build on-disk caches of precomputed workload
colbatches for testing/benchmarking. If we write each batch in some
workload.Table as a record in the file, the random access will allow us
to later fake out a workload.Table to use in the test that just fills
the precomputed data from a mmap'd file.

The Reset call I added to ArrowToBatch will help with the above, but
seems to have slowed the benchmarks down a bit, though the diff is small
in absolute terms. Reset is clearing the nulls just for ArrowToBatch to
then go and overwrite them, which is wasteful. Avoiding the waste is a
bit awkward but I think gets us most of this back if we care.

    name                                                        old time/op    new time/op    delta
    ArrowBatchConverter/Bool/nullFraction=0.00/ArrowToBatch-8     2.02µs ± 3%    2.22µs ± 7%   +9.99%  (p=0.008 n=5+5)
    ArrowBatchConverter/Bytes/nullFraction=0.00/ArrowToBatch-8    2.83µs ± 1%    3.13µs ± 9%  +10.41%  (p=0.008 n=5+5)
    ArrowBatchConverter/Int64/nullFraction=0.00/ArrowToBatch-8     145ns ± 1%     173ns ± 4%  +19.36%  (p=0.008 n=5+5)

    name                                                        old speed      new speed      delta
    ArrowBatchConverter/Bool/nullFraction=0.00/ArrowToBatch-8    507MB/s ± 3%   462MB/s ± 7%   -8.96%  (p=0.008 n=5+5)
    ArrowBatchConverter/Bytes/nullFraction=0.00/ArrowToBatch-8  23.1GB/s ± 1%  21.0GB/s ± 8%   -9.09%  (p=0.008 n=5+5)
    ArrowBatchConverter/Int64/nullFraction=0.00/ArrowToBatch-8  56.4GB/s ± 1%  47.3GB/s ± 4%  -16.09%  (p=0.008 n=5+5)

    name                                                        old alloc/op   new alloc/op   delta
    ArrowBatchConverter/Bool/nullFraction=0.00/ArrowToBatch-8      64.0B ± 0%     64.0B ± 0%     ~     (all equal)
    ArrowBatchConverter/Bytes/nullFraction=0.00/ArrowToBatch-8     96.0B ± 0%     96.0B ± 0%     ~     (all equal)
    ArrowBatchConverter/Int64/nullFraction=0.00/ArrowToBatch-8     96.0B ± 0%     96.0B ± 0%     ~     (all equal)

    name                                                        old allocs/op  new allocs/op  delta
    ArrowBatchConverter/Bool/nullFraction=0.00/ArrowToBatch-8       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
    ArrowBatchConverter/Bytes/nullFraction=0.00/ArrowToBatch-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
    ArrowBatchConverter/Int64/nullFraction=0.00/ArrowToBatch-8      2.00 ± 0%      2.00 ± 0%     ~     (all equal)

Release note: None