@danhhz 2019-06-11
I was using this in a followup PR and needed to parse the types back out of the schema, which uncovered a couple bugs with this (geez flatbuffers is hard to use correctly). Figured I may as well pull in the fixes and tests now