@danhhz 2019-07-10
Gonna go ahead and merge this before it rots again. The mmap stuff will be a followup now so if either of you end up looking at this again and have any final comments, it'll be easy to sneak them into that followup

bors r=tbg,asubiotto