@tbg 2019-07-10
Sgtm

On Wed, Jul 10, 2019, 19:44 Daniel Harrison <notifications@github.com>
wrote:

> Gonna go ahead and merge this before it rots again. The mmap stuff will be
> a followup now so if either of you end up looking at this again and have
> any final comments, it'll be easy to sneak them into that followup
>
> bors r=tbg,asubiotto
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/37803?email_source=notifications&email_token=ABGXPZBBQEUZWLR6BDJ7HOTP6YNWVA5CNFSM4HPR4UO2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODZUG3QA#issuecomment-510160320>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABGXPZFHQ5X56TXKO53IYC3P6YNWVANCNFSM4HPR4UOQ>
> .
>